refactor: move analytics for a hook
